refactor: PatternDisplay からフルスクリーン/読込フックと PatternCanvas を抽出 (#14)
